# Make 'for' loop in dsiReallocate parallel