As noted in JIRA 189 and a comment in the code, our dsiReallocate()
code contained a serial for loop that by all rights should have been
parallel.  The last time I tried to convert it over, I had failures in
testing, but those failures were addressed by @mppf in 5f6c815,
and this seems to be working now, so great!