@ronawho 2018-06-16
FYI it looks like Michael fixed the underlying issue in 5f6c8150483 (part of the massive array rework). The problem before was the `followMe(i).low*rStride`, which was a `uint*int` for something like RA.

I'd still be curious to learn why we're calling dsiReallocate in RA though