@bradcray 2018-06-18
This passed standard testing, but I think I'm going to wait to merge it until after my enum range rework goes in.