@bradcray 2018-06-18
@ronawho: Thinking ahead, do you think we'd want to perf playground this before merging?