@bradcray 2018-06-19
@ronawho: The dsiReallocate in RA is due to resizing the targetLocales domain and array in the Block distribution when setting everything up.

@mppf: Alerting you to this PR, as you'd added a comment to the source pointing out the serial for loop at some point int he past.

This will be tested in the perf playground tonight.