@ronawho 2018-06-21
https://chapel-lang.org/perf/chapcs/dsiReallocPar/?startdate=2018/06/08&enddate=2018/06/29&graphs=arrayvectoroperations,compilationtime,jacobiemittedcodesize

Slight increase in compilation time and emitted code size, but nice improvements for array-as-vec operations:
 - push_back went from ~0.65s => ~0.50s
 - pop_front went from ~0.40s => ~0.30s
 - and improvements for most other operations, but I think those are the most common ones